requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
numpy==2.1.3
python-telegram-bot==21.9
python-dotenv==1.0.1
requests==2.32.3
